requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
holidays==0.49
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
numpy==1.26.4
openai==1.30.5
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
prophet==1.1.5
protobuf==4.25.3
pyarrow==16.1.0
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stanio==0.5.0
streamlit==1.35.0
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
